+++ logprep/run_logprep.py
@@ -10,7 +10,6 @@ import warnings
 import click
 
 from logprep.generator.factory import ControllerFactory
-from logprep.generator.kafka.run_load_tester import LoadTester
 from logprep.runner import Runner
 from logprep.util.ansi import Fore
 from logprep.util.auto_rule_tester.auto_rule_tester import AutoRuleTester
@@ -247,23 +246,8 @@ def generate():
     """
 
 
-@generate.command(name="kafka")
-@click.argument("config")
-@click.option(
-    "--file", help="Path to file with documents", default=None, type=click.Path(exists=True)
-)
-def generate_kafka(config, file):
-    """
-    Generate events by taking them from kafka or a jsonl file and sending them to Kafka.
-
-    CONFIG is a path to a configuration file for the event generation.
-    """
-    load_tester = LoadTester(config, file)
-    load_tester.run()
-
-
 @common_generate_options
-@generate.command(name="kafka2")
+@generate.command(name="kafka")
 @click.option(
     "--output-config",
     help="Enables http generator to use kafka output",
@@ -277,7 +261,7 @@ def generate_kafka(config, file):
     default=0,
     type=int,
 )
-def generate_kafka2(**kwargs):
+def generate_kafka(**kwargs):
     """
     Generates events based on templated sample files stored inside a dataset directory.
     The events will be sent to a kafka endpoint.